# Conflicts:
#	source/frontend-user/lib/features/profile/presentation/screens/edit_profile_screen.dart

+++ source/BE/ecommerceproject/ecommerceproject/src/main/java/com/example/ecommerceproject/model/Coupon.java
@@ -21,7 +21,7 @@ public class Coupon {
 
     private String code;
 
-    private int value;
+    private double value;
 
     private int maxUses;
 
@@ -32,12 +32,14 @@ public class Coupon {
     private List<String> ordersApplied = new ArrayList<>();
 
     // Constructor với các tham số cơ bản
-    public Coupon(String code, int value, int maxUses) {
+    public Coupon(String code, double value, int maxUses) {
         this.code = code;
         this.value = value;
         this.maxUses = maxUses;
         this.usedCount = 0;
         this.creationTime = LocalDateTime.now();
+
+        
     }
 
     // Phương thức kiểm tra xem coupon có thể sử dụng được không
